Enable AT builds on RHEL8
Building on systems with optimized libraries can conflict with the
AT build.  The loader, including the loader that is built with AT,
will prefer libraries in so-called "optimized" paths.  These are
any paths which are subdirectories of a path in the `ld.so.conf`
configuration which are classified as "optimized" by a predefined
list, generally by processor architecture, like `power9`.
So, a file in `/lib/power9/libc.so` will take preference over
`/opt/atX/lib/libc.so`, even if `/opt/atX/lib` is earlier than
`/lib` in the `ld.so.conf` search order.

The problem coveniently occurs between the two `ldconfig`
stages.  Work around the problem by creating a fake "optimized"
set of libraries which are simply copied from the "unoptimized"
directory into "optimized" subdirectories of a new (temporary)
directory, `slib-override`.

This new directory is removed from `ld.so.conf` and the directory
tree is completely removed before the second `ldconfig` runs.

Fixes #1298.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>